fix: Round all the things
* Convert all column in summary that store quantity or units to integer as they should never have decimals
* Update summary calculation to round unit penalties

+++ backend/lcfs/db/models/compliance/ComplianceReportSummary.py
@@ -2,7 +2,6 @@ from sqlalchemy import Column, Integer, Float, ForeignKey, Boolean, CheckConstra
 from sqlalchemy.exc import InvalidRequestError
 from sqlalchemy.orm import relationship
 from lcfs.db.base import BaseModel, Auditable
-from datetime import datetime
 
 
 class ComplianceReportSummary(BaseModel, Auditable):
@@ -19,70 +18,78 @@ class ComplianceReportSummary(BaseModel, Auditable):
     is_locked = Column(Boolean, default=False)
 
     # Renewable fuel target summary
-    line_1_fossil_derived_base_fuel_gasoline = Column(Float, nullable=False, default=0)
-    line_1_fossil_derived_base_fuel_diesel = Column(Float, nullable=False, default=0)
-    line_1_fossil_derived_base_fuel_jet_fuel = Column(Float, nullable=False, default=0)
+    line_1_fossil_derived_base_fuel_gasoline = Column(
+        Integer, nullable=False, default=0
+    )
+    line_1_fossil_derived_base_fuel_diesel = Column(Integer, nullable=False, default=0)
+    line_1_fossil_derived_base_fuel_jet_fuel = Column(
+        Integer, nullable=False, default=0
+    )
 
     line_2_eligible_renewable_fuel_supplied_gasoline = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
     line_2_eligible_renewable_fuel_supplied_diesel = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
     line_2_eligible_renewable_fuel_supplied_jet_fuel = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
 
     line_3_total_tracked_fuel_supplied_gasoline = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
-    line_3_total_tracked_fuel_supplied_diesel = Column(Float, nullable=False, default=0)
+    line_3_total_tracked_fuel_supplied_diesel = Column(
+        Integer, nullable=False, default=0
+    )
     line_3_total_tracked_fuel_supplied_jet_fuel = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
 
     line_4_eligible_renewable_fuel_required_gasoline = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
     line_4_eligible_renewable_fuel_required_diesel = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
     line_4_eligible_renewable_fuel_required_jet_fuel = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
 
     line_5_net_notionally_transferred_gasoline = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
-    line_5_net_notionally_transferred_diesel = Column(Float, nullable=False, default=0)
+    line_5_net_notionally_transferred_diesel = Column(
+        Integer, nullable=False, default=0
+    )
     line_5_net_notionally_transferred_jet_fuel = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
 
-    line_6_renewable_fuel_retained_gasoline = Column(Float, nullable=False, default=0)
-    line_6_renewable_fuel_retained_diesel = Column(Float, nullable=False, default=0)
-    line_6_renewable_fuel_retained_jet_fuel = Column(Float, nullable=False, default=0)
+    line_6_renewable_fuel_retained_gasoline = Column(Integer, nullable=False, default=0)
+    line_6_renewable_fuel_retained_diesel = Column(Integer, nullable=False, default=0)
+    line_6_renewable_fuel_retained_jet_fuel = Column(Integer, nullable=False, default=0)
 
-    line_7_previously_retained_gasoline = Column(Float, nullable=False, default=0)
-    line_7_previously_retained_diesel = Column(Float, nullable=False, default=0)
-    line_7_previously_retained_jet_fuel = Column(Float, nullable=False, default=0)
+    line_7_previously_retained_gasoline = Column(Integer, nullable=False, default=0)
+    line_7_previously_retained_diesel = Column(Integer, nullable=False, default=0)
+    line_7_previously_retained_jet_fuel = Column(Integer, nullable=False, default=0)
 
-    line_8_obligation_deferred_gasoline = Column(Float, nullable=False, default=0)
-    line_8_obligation_deferred_diesel = Column(Float, nullable=False, default=0)
-    line_8_obligation_deferred_jet_fuel = Column(Float, nullable=False, default=0)
+    line_8_obligation_deferred_gasoline = Column(Integer, nullable=False, default=0)
+    line_8_obligation_deferred_diesel = Column(Integer, nullable=False, default=0)
+    line_8_obligation_deferred_jet_fuel = Column(Integer, nullable=False, default=0)
 
-    line_9_obligation_added_gasoline = Column(Float, nullable=False, default=0)
-    line_9_obligation_added_diesel = Column(Float, nullable=False, default=0)
-    line_9_obligation_added_jet_fuel = Column(Float, nullable=False, default=0)
+    line_9_obligation_added_gasoline = Column(Integer, nullable=False, default=0)
+    line_9_obligation_added_diesel = Column(Integer, nullable=False, default=0)
+    line_9_obligation_added_jet_fuel = Column(Integer, nullable=False, default=0)
 
     line_10_net_renewable_fuel_supplied_gasoline = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
     line_10_net_renewable_fuel_supplied_diesel = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
     line_10_net_renewable_fuel_supplied_jet_fuel = Column(
-        Float, nullable=False, default=0
+        Integer, nullable=False, default=0
     )
 
     line_11_non_compliance_penalty_gasoline = Column(Float, nullable=True, default=0)
@@ -90,17 +97,17 @@ class ComplianceReportSummary(BaseModel, Auditable):
     line_11_non_compliance_penalty_jet_fuel = Column(Float, nullable=True, default=0)
 
     # Low carbon fuel target summary columns
-    line_12_low_carbon_fuel_required = Column(Float, nullable=False, default=0)
-    line_13_low_carbon_fuel_supplied = Column(Float, nullable=False, default=0)
-    line_14_low_carbon_fuel_surplus = Column(Float, nullable=False, default=0)
-    line_15_banked_units_used = Column(Float, nullable=False, default=0)
-    line_16_banked_units_remaining = Column(Float, nullable=False, default=0)
-    line_17_non_banked_units_used = Column(Float, nullable=False, default=0)
-    line_18_units_to_be_banked = Column(Float, nullable=False, default=0)
-    line_19_units_to_be_exported = Column(Float, nullable=False, default=0)
-    line_20_surplus_deficit_units = Column(Float, nullable=False, default=0)
+    line_12_low_carbon_fuel_required = Column(Integer, nullable=False, default=0)
+    line_13_low_carbon_fuel_supplied = Column(Integer, nullable=False, default=0)
+    line_14_low_carbon_fuel_surplus = Column(Integer, nullable=False, default=0)
+    line_15_banked_units_used = Column(Integer, nullable=False, default=0)
+    line_16_banked_units_remaining = Column(Integer, nullable=False, default=0)
+    line_17_non_banked_units_used = Column(Integer, nullable=False, default=0)
+    line_18_units_to_be_banked = Column(Integer, nullable=False, default=0)
+    line_19_units_to_be_exported = Column(Integer, nullable=False, default=0)
+    line_20_surplus_deficit_units = Column(Integer, nullable=False, default=0)
     line_21_surplus_deficit_ratio = Column(Float, nullable=False, default=0)
-    line_22_compliance_units_issued = Column(Float, nullable=False, default=0)
+    line_22_compliance_units_issued = Column(Integer, nullable=False, default=0)
 
     # Non-compliance penalty summary columns
     line_11_fossil_derived_base_fuel_gasoline = Column(Float, nullable=False, default=0)
